refactor: address PR #129 review feedback
Threads addressed:

  - "completes the OTP flow" / "do we already have a step for this?"
    (e2e/step-definitions/session-reuse-bugs.steps.ts:628,650): the
    monolithic step duplicated the existing canonical sequence used
    elsewhere in the suite. Replace it in row 5 with the established
    composition: "enters the test email" → "OTP email arrives" →
    "enters the OTP code", and drop the bespoke step body entirely.

  - "this is bloating main()" (packages/pds-core/src/index.ts:1102):
    extract the EPDS_TEST_HOOKS=1 wiring (the expire-device-account
    route + its production-mode guard + the secret check) into a new
    lib/test-hooks.ts. main() now installs it via a single
    installTestHooks() call, matching the pattern used for the other
    middleware modules.

  - "guard is now misnamed" (welcome-page-guard.ts:248): expand the
    file's top docstring to describe both UIs the guard now suppresses
    (welcome page rows 1–4; sign-in-view rows 5/6/9) and explain why
    the filename is kept stable. The function/file names stay as-is to
    avoid churning every import site for no behavioural gain — the
    docstring carries the current scope.

The fourth thread (welcome-page-guard.ts:281, "isn't replicating
upstream's render decision brittle?") is a real architectural
question; left for separate discussion in the PR rather than papered
over in code.
