+++ packages/pds-core/src/index.ts
@@ -64,6 +64,7 @@ import { createChooserEnrichmentMiddleware } from './chooser-enrichment.js'
 import { createUpstreamFaviconMiddleware } from './upstream-favicon.js'
 import { createWelcomePageGuard } from './welcome-page-guard.js'
 import { loadDeviceAccountEmails } from './lib/device-accounts.js'
+import { installTestHooks } from './lib/test-hooks.js'
 
 const logger = createLogger('pds-core')
 
@@ -1088,79 +1089,7 @@ async function main() {
     res.json({ emails })
   })
 
-  // =========================================================================
-  // Test-only hooks for the e2e suite. Mounted only when EPDS_TEST_HOOKS=1
-  // and refused outright when NODE_ENV=production. Mirrors auth-service's
-  // /_internal/test/expire-otp pattern: a narrow UPDATE that backdates a
-  // single timestamp to reproduce time-dependent behaviour without waiting
-  // out the wall-clock TTL. Used by features/session-reuse-bugs.feature
-  // rows 6 and 9 to age account_device.updatedAt past upstream's
-  // authenticationMaxAge (7 days) so checkLoginRequired returns true for
-  // the targeted binding(s).
-  // =========================================================================
-
-  if (process.env.EPDS_TEST_HOOKS === '1') {
-    if (process.env.NODE_ENV === 'production') {
-      throw new Error(
-        'EPDS_TEST_HOOKS=1 is set but NODE_ENV=production — refusing to mount test-only endpoints',
-      )
-    }
-    logger.warn(
-      'Test hooks ENABLED — /_internal/test/* routes are live (EPDS_TEST_HOOKS=1)',
-    )
-
-    pds.app.post(
-      '/_internal/test/expire-device-account',
-      express.json(),
-      async (req, res) => {
-        if (!verifyInternalSecret(req.headers['x-internal-secret'])) {
-          res.status(401).json({ error: 'Unauthorized' })
-          return
-        }
-        const did = ((req.body?.did as string) || '').trim()
-        const deviceId =
-          typeof req.body?.deviceId === 'string'
-            ? req.body.deviceId.trim()
-            : undefined
-        if (!did) {
-          res.status(400).json({ error: 'Missing did' })
-          return
-        }
-        // 8 days ago — comfortably past upstream's 7-day authenticationMaxAge
-        // so checkLoginRequired returns true on every backdated row.
-        const past = new Date(
-          Date.now() - 8 * 24 * 60 * 60 * 1000,
-        ).toISOString()
-        try {
-          // Reuse the PDS accountManager's own Kysely instance — same handle
-          // PDS uses for upsertDeviceAccount, so there are no two-connection
-          // WAL-visibility surprises.
-          // eslint-disable-next-line @typescript-eslint/no-explicit-any -- account_device shape not exported by @atproto/pds
-          const db = pds.ctx.accountManager.db.db as any
-          let q = db
-            .updateTable('account_device')
-            .set({ updatedAt: past })
-            .where('did', '=', did)
-          if (deviceId) {
-            q = q.where('deviceId', '=', deviceId)
-          }
-          const result = await q.executeTakeFirst()
-          const updated = Number(result?.numUpdatedRows ?? 0)
-          logger.warn(
-            { did, deviceId, updated, past },
-            'Backdated account_device.updatedAt',
-          )
-          res.json({ updated })
-        } catch (err) {
-          logger.error(
-            { err, did, deviceId },
-            'Failed to backdate account_device.updatedAt',
-          )
-          res.status(500).json({ error: 'Internal server error' })
-        }
-      },
-    )
-  }
+  installTestHooks({ pds, app: pds.app, logger })
 
   // =========================================================================
   // TLS check - used by Caddy on-demand TLS to verify handle ownership
